↺ merge  svn merge -r 1544:1694 https://amethyst.yui.ne.jp/svn/pc98/np2/trunk  svn merge -r 1695:1697 https://amethyst.yui.ne.jp/svn/pc98/np2/trunk  svn merge -r 1698:1734 https://amethyst.yui.ne.jp/svn/pc98/np2/trunk
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2529 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ support_record_avi/codecnv/sjiseuc.c
@@ -39,7 +39,7 @@ UINT codecnv_sjistoeuc(char *lpOutput, UINT cchOutput, const char *lpInput, UINT
 	else
 	{
 		// String mode
-		nLength = sjistoeuc(lpOutput, cchOutput - 1, lpInput, strlen(lpInput));
+		nLength = sjistoeuc(lpOutput, cchOutput - 1, lpInput, (UINT)strlen(lpInput));
 		if (lpOutput)
 		{
 			lpOutput[nLength] = '\0';
